mcp/alibabacloud-lindorm-mcp-server/pyproject.toml
[project]
name = "alibabacloud-lindorm-mcp-server"
version = "0.1.0"
description = "alibaba cloud lindorm mcp-server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "mysql-connector-python==8.0.11",
    "opensearch-py>=2.8.0",
    "pytest>=8.3.5",
]
